+++ src/ControlTasks/__init__.py
@@ -2,8 +2,9 @@ from .clock_manager import ClockManager
 from .control_task_base import ControlTaskBase
 from .mission_manager import MissionManager
 from .test_control_task import TestControlTask
-from .read_camera import ReadCamera
+from .read_webcam import Webcam
 from .point_tracker import PointTracker
 from .process_frame import ProcessFrame
 from .display_frame import DisplayFrame
-from .read_full_camera import ReadFullCamera
+from .read_depth_camera import DepthCamera
+from .read_static_frame import StaticFrame
