+++ src/ControlTasks/read_static_frame.py
