+++ src/ControlTasks/read_webcam.py
@@ -2,8 +2,7 @@ from .control_task_base import ControlTaskBase
 import cv2
 
 
-class ReadCamera(ControlTaskBase):
-
+class Webcam(ControlTaskBase):
     def setup(self):
         self.vc = cv2.VideoCapture(0)
         if self.vc.isOpened():
@@ -11,7 +10,7 @@ class ReadCamera(ControlTaskBase):
             rval, frame = self.vc.read()
 
         # to make processing code happy
-        self.execute() #TODO REMOVE
+        self.execute()  #TODO REMOVE
 
     def default(self):
         self.sfr.set("curr_frame", None)
