+++ src/__init__.py
@@ -1,12 +1,3 @@
 '''__init__.py for src'''
 
-from .ControlTasks import (
-    ClockManager,
-    ControlTaskBase,
-    MissionManager,
-    PointTracker,
-    ReadCamera,
-    TestControlTask,
-)
-
 from . import ControlTasks, framework, runner, sfr, vision
